Refactor: split out classes from QueryCompiler.scala
